requirements-dev.txt
-r requirements.txt
pytest>=8.0
pytest-django>=4.9
pytest-describe>=2.2
pytest-leela>=0.1
pytest-cov>=6.0
ruff>=0.8
mypy>=1.13
django-stubs>=5.1
factory-boy>=3.3
faker>=30.0
